代码内增加注释 更新带初始验证的私聊模块 新的demo 大量代码更新 2009.11.17 by X

+++  xsimplechat/SimpleChat/src/model/SendNode.java
@@ -2,18 +2,18 @@ package model;
 
 public class SendNode {
 	private byte[] msg;
-	private int sid;
+	private long sid;
 
 	public SendNode() {
 		this(0, new byte[0]);
 	}
 
-	public SendNode(int sid, byte[] msg) {
+	public SendNode(long sid, byte[] msg) {
 		this.sid = sid;
 		this.msg = msg;
 	}
 
-	public int getSid() {
+	public long getSid() {
 		return sid;
 	}
